// src/api/generated/models/cartItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Kuppams Organic E-commerce API
 * A comprehensive REST API for an organic products e-commerce platform
 * OpenAPI spec version: 1.0.0
 */
import type { Product } from './product';

export interface CartItem {
  product?: Product;
  quantity?: number;
}